Moving to general category as suggested in review
